SignInStatuses.cs (small)

+++ Assets/Scripts/Server/PlayFab/Unattachable/SignInStatuses.cs
@@ -1,6 +1,6 @@
 namespace Server.General {
 	internal static class SignInStatuses: object {
-		internal enum SignInStatus: byte {
+		internal enum SignInStatus: int {
 			None,
 			NoUsernameOrEmail,
 			NoPassword,
